add copy-button for code blocks
To add copy button for code-blocks.

note1:
Before compiling on a local machine using 'make html' the 'sphinx_copybutton' package using (pip install sphinx_copybutton) should be installed.

note2:
for workflow it is needed to add the following to Sphinx.yml  (in jobs>steps). But I didn't do it because maybe you are doing it using another way!

"
# Install dependencies
    - name: Install dependencies
      run: |
          python -m pip install --upgrade pip
          pip install -r requirements.txt
"